Update to version 0.9.8

+++ ApplicationCode/ProjectDataModel/RimUiTreeView.cpp
@@ -35,6 +35,7 @@
 #include "RigCaseCellResultsData.h"
 #include "RimStatisticsCase.h"
 #include "RimResultCase.h"
+#include "RimMimeData.h"
 
 //--------------------------------------------------------------------------------------------------
 /// 
@@ -1097,7 +1098,7 @@ bool RimUiTreeView::userConfirmedGridCaseGroupChange(const QModelIndexList& item
             else
             {
                 questionText = "This operation will invalidate statistics results in grid case groups\n";
-                for (int i = 0; i < typedObjects.size(); i++)
+                for (size_t i = 0; i < typedObjects.size(); i++)
                 {
                     questionText += QString("\"%1\"\n").arg(typedObjects[i]->name());
                 }
